Condensed all mod files into one class and moved it to the test directory
