--- mdk/src/main/java/com/example/examplemod/datagen/ModBlockStateProvider.java
@@ -1,24 +1,0 @@
-package com.example.examplemod.datagen;
-
-import com.example.examplemod.ExampleMod;
-import com.example.examplemod.block.ModBlocks;
-import net.minecraft.data.PackOutput;
-import net.minecraft.world.level.block.Block;
-import net.minecraftforge.client.model.generators.BlockStateProvider;
-import net.minecraftforge.common.data.ExistingFileHelper;
-import net.minecraftforge.registries.RegistryObject;
-
-public class ModBlockStateProvider extends BlockStateProvider {
-    public ModBlockStateProvider(PackOutput output, ExistingFileHelper exFileHelper) {
-        super(output, ExampleMod.MODID, exFileHelper);
-    }
-
-    @Override
-    protected void registerStatesAndModels() {
-        blockWithItem(ModBlocks.SILVER_ORE);
-    }
-
-    private void blockWithItem(RegistryObject<Block> blockRegistryObject) {
-        simpleBlockWithItem(blockRegistryObject.get(), cubeAll(blockRegistryObject.get()));
-    }
-}
