--- mdk/src/main/java/com/example/examplemod/datagen/loot/ModBlockLootTables.java
@@ -1,25 +1,0 @@
-package com.example.examplemod.datagen.loot;
-
-import com.example.examplemod.block.ModBlocks;
-import net.minecraft.data.loot.BlockLootSubProvider;
-import net.minecraft.world.flag.FeatureFlags;
-import net.minecraft.world.level.block.Block;
-import net.minecraftforge.registries.RegistryObject;
-
-import java.util.Set;
-
-public class ModBlockLootTables extends BlockLootSubProvider {
-    public ModBlockLootTables() {
-        super(Set.of(), FeatureFlags.REGISTRY.allFlags());
-    }
-
-    @Override
-    protected void generate() {
-        this.dropSelf(ModBlocks.SILVER_ORE.get());
-    }
-
-    @Override
-    protected Iterable<Block> getKnownBlocks() {
-        return ModBlocks.BLOCKS.getEntries().stream().map(RegistryObject::get)::iterator;
-    }
-}
