--- mdk/src/main/java/com/example/examplemod/item/ModItems.java
@@ -1,16 +1,0 @@
-package com.example.examplemod.item;
-
-import com.example.examplemod.ExampleMod;
-import net.minecraft.world.item.Item;
-import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.registries.DeferredRegister;
-import net.minecraftforge.registries.ForgeRegistries;
-import net.minecraftforge.registries.RegistryObject;
-
-public class ModItems {
-    public static final DeferredRegister<Item> ITEMS =
-            DeferredRegister.create(ForgeRegistries.ITEMS, ExampleMod.MODID);
-    public static void register(IEventBus eventBus) {
-        ITEMS.register(eventBus);
-    }
-}
