--- mdk/src/main/java/com/example/examplemod/worldgen/ModOrePlacement.java
@@ -1,15 +1,0 @@
-package com.example.examplemod.worldgen;
-
-import net.minecraft.world.level.levelgen.placement.*;
-
-import java.util.List;
-
-public class ModOrePlacement {
-    public static List<PlacementModifier> orePlacement(PlacementModifier m1, PlacementModifier m2) {
-        return List.of(m1, InSquarePlacement.spread(), m2, BiomeFilter.biome());
-    }
-
-    public static List<PlacementModifier> commonOrePlacement(int count, PlacementModifier heightRange) {
-        return orePlacement(CountPlacement.of(count), heightRange);
-    }
-}
