--- mdk/src/main/resources/pack.mcmeta
@@ -1,6 +1,0 @@
-{
-  "pack": {
-    "description": "${mod_id} resources",
-    "pack_format": 32
-  }
-}
