+++ src/test/java/net/minecraftforge/debug/world/BiomeDecorationTest.java
@@ -1,0 +1,262 @@
+package net.minecraftforge.debug.world;
+
+import com.mojang.logging.LogUtils;
+import java.util.List;
+import java.util.Set;
+import java.util.concurrent.CompletableFuture;
+import java.util.function.Supplier;
+import net.minecraft.core.Holder;
+import net.minecraft.core.HolderGetter;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.core.HolderSet;
+import net.minecraft.core.RegistrySetBuilder;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.data.DataGenerator;
+import net.minecraft.data.PackOutput;
+import net.minecraft.data.loot.BlockLootSubProvider;
+import net.minecraft.data.loot.LootTableProvider;
+import net.minecraft.data.worldgen.BootstrapContext;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.BiomeTags;
+import net.minecraft.tags.BlockTags;
+import net.minecraft.world.flag.FeatureFlags;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.levelgen.GenerationStep;
+import net.minecraft.world.level.levelgen.VerticalAnchor;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
+import net.minecraft.world.level.levelgen.feature.Feature;
+import net.minecraft.world.level.levelgen.feature.configurations.FeatureConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.OreConfiguration;
+import net.minecraft.world.level.levelgen.placement.BiomeFilter;
+import net.minecraft.world.level.levelgen.placement.CountPlacement;
+import net.minecraft.world.level.levelgen.placement.HeightRangePlacement;
+import net.minecraft.world.level.levelgen.placement.InSquarePlacement;
+import net.minecraft.world.level.levelgen.placement.PlacedFeature;
+import net.minecraft.world.level.levelgen.placement.PlacementModifier;
+import net.minecraft.world.level.levelgen.structure.templatesystem.RuleTest;
+import net.minecraft.world.level.levelgen.structure.templatesystem.TagMatchTest;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.model.generators.BlockStateProvider;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.common.data.DatapackBuiltinEntriesProvider;
+import net.minecraftforge.common.data.ExistingFileHelper;
+import net.minecraftforge.common.world.BiomeModifier;
+import net.minecraftforge.common.world.ForgeBiomeModifiers;
+import net.minecraftforge.data.event.GatherDataEvent;
+import net.minecraftforge.event.BuildCreativeModeTabContentsEvent;
+import net.minecraftforge.event.server.ServerStartingEvent;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
+import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.registries.DeferredRegister;
+import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.registries.RegistryObject;
+import org.slf4j.Logger;
+
+// The value here should match an entry in the META-INF/mods.toml file
+@Mod(BiomeDecorationTest.MODID)
+public class BiomeDecorationTest {
+    public static final String MODID = "examplemod";
+    private static final Logger LOGGER = LogUtils.getLogger();
+
+    public BiomeDecorationTest() {
+        IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
+        ModItems.register(modEventBus);
+        ModBlocks.register(modEventBus);
+        modEventBus.addListener(this::commonSetup);
+        MinecraftForge.EVENT_BUS.register(this);
+        modEventBus.addListener(this::addCreative);
+    }
+
+    private void commonSetup(final FMLCommonSetupEvent event) {
+
+    }
+
+    // Add the example block item to the building blocks tab
+    private void addCreative(BuildCreativeModeTabContentsEvent event) {
+
+    }
+
+    // You can use SubscribeEvent and let the Event Bus discover methods to call
+    @SubscribeEvent
+    public void onServerStarting(ServerStartingEvent event) {
+
+    }
+
+    // You can use EventBusSubscriber to automatically register all static methods in the class annotated with @SubscribeEvent
+    @Mod.EventBusSubscriber(modid = MODID, bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
+    public static class ClientModEvents {
+        @SubscribeEvent
+        public static void onClientSetup(FMLClientSetupEvent event) {
+
+        }
+    }
+
+    public class ModBlocks {
+        public static final DeferredRegister<Block> BLOCKS =
+                DeferredRegister.create(ForgeRegistries.BLOCKS, BiomeDecorationTest.MODID);
+        public static final RegistryObject<Block> SILVER_ORE = registerBlock("silver_ore",
+                () -> new Block(BlockBehaviour.Properties.ofFullCopy(Blocks.GOLD_ORE)));
+
+        private static <T extends Block> RegistryObject<T> registerBlock(String name, Supplier<T> block) {
+            RegistryObject<T> toReturn = BLOCKS.register(name, block);
+            registerBlockItem(name, toReturn);
+            return toReturn;
+        }
+
+        private static <T extends Block> RegistryObject<Item> registerBlockItem(String name, RegistryObject<T> block) {
+            return ModItems.ITEMS.register(name, () -> new BlockItem(block.get(), new Item.Properties()));
+        }
+
+        public static void register(IEventBus eventBus) {
+            BLOCKS.register(eventBus);
+        }
+    }
+
+    public class ModItems {
+        public static final DeferredRegister<Item> ITEMS =
+                DeferredRegister.create(ForgeRegistries.ITEMS, BiomeDecorationTest.MODID);
+
+        public static void register(IEventBus eventBus) {
+            ITEMS.register(eventBus);
+        }
+    }
+
+    public class ModBiomeModifiers {
+
+        public static final ResourceKey<BiomeModifier> ADD_OVERWORLD_SILVER_ORE = registerKey("add_overworld_silver_ore");
+
+        public static void bootstrap(BootstrapContext<BiomeModifier> context) {
+            var placedFeatures = context.lookup(Registries.PLACED_FEATURE);
+            var biomes = context.lookup(Registries.BIOME);
+
+            context.register(ADD_OVERWORLD_SILVER_ORE, new ForgeBiomeModifiers.AddFeaturesBiomeModifier(
+                    biomes.getOrThrow(BiomeTags.IS_OVERWORLD),
+                    HolderSet.direct(placedFeatures.getOrThrow(ModPlacedFeatures.OVERWORLD_SILVER_ORE_PLACED_KEY)),
+                    GenerationStep.Decoration.UNDERGROUND_ORES));
+        }
+
+        private static ResourceKey<BiomeModifier> registerKey(String name) {
+            return ResourceKey.create(ForgeRegistries.Keys.BIOME_MODIFIERS, new ResourceLocation(BiomeDecorationTest.MODID, name));
+        }
+    }
+
+    public class ModConfiguredFeatures {
+
+        public static final ResourceKey<ConfiguredFeature<?, ?>> OVERWORLD_SILVER_ORE_KEY = registerKey("overworld_silver_ore");
+
+        public static void bootstrap(BootstrapContext<ConfiguredFeature<?, ?>> context) {
+            RuleTest stoneReplaceables = new TagMatchTest(BlockTags.STONE_ORE_REPLACEABLES);
+
+            List<OreConfiguration.TargetBlockState> overworldSilverOres = List.of(OreConfiguration.target(stoneReplaceables,
+                    ModBlocks.SILVER_ORE.get().defaultBlockState()));
+
+            register(context, OVERWORLD_SILVER_ORE_KEY, Feature.ORE, new OreConfiguration(overworldSilverOres, 9));
+
+        }
+
+        public static ResourceKey<ConfiguredFeature<?, ?>> registerKey(String name) {
+            return ResourceKey.create(Registries.CONFIGURED_FEATURE, new ResourceLocation(BiomeDecorationTest.MODID, name));
+        }
+
+        private static <FC extends FeatureConfiguration, F extends Feature<FC>> void register(BootstrapContext<ConfiguredFeature<?, ?>> context, ResourceKey<ConfiguredFeature<?, ?>> key, F feature, FC configuration) {
+            context.register(key, new ConfiguredFeature<>(feature, configuration));
+        }
+    }
+
+    public class ModPlacedFeatures {
+
+        public static final ResourceKey<PlacedFeature> OVERWORLD_SILVER_ORE_PLACED_KEY = registerKey("overworld_silver_ore_placed");
+
+        public static void bootstrap(BootstrapContext<PlacedFeature> context) {
+            HolderGetter<ConfiguredFeature<?, ?>> configuredFeatures = context.lookup(Registries.CONFIGURED_FEATURE);
+
+            register(context, OVERWORLD_SILVER_ORE_PLACED_KEY, configuredFeatures.getOrThrow(ModConfiguredFeatures.OVERWORLD_SILVER_ORE_KEY),
+                    List.of(CountPlacement.of(4), InSquarePlacement.spread(), HeightRangePlacement.triangle(VerticalAnchor.absolute(-64), VerticalAnchor.absolute(32)), BiomeFilter.biome()));
+        }
+
+        private static ResourceKey<PlacedFeature> registerKey(String name) {
+            return ResourceKey.create(Registries.PLACED_FEATURE, new ResourceLocation(BiomeDecorationTest.MODID, name));
+        }
+
+        private static void register(BootstrapContext<PlacedFeature> context, ResourceKey<PlacedFeature> key, Holder<ConfiguredFeature<?, ?>> configuration, List<PlacementModifier> modifiers) {
+            context.register(key, new PlacedFeature(configuration, List.copyOf(modifiers)));
+        }
+    }
+
+    @Mod.EventBusSubscriber(modid = BiomeDecorationTest.MODID, bus = Mod.EventBusSubscriber.Bus.MOD)
+    public class DataGenerators {
+        @SubscribeEvent
+        public static void gatherdata(GatherDataEvent event) {
+            DataGenerator generator = event.getGenerator();
+            PackOutput packOutput = generator.getPackOutput();
+            ExistingFileHelper existingFileHelper = event.getExistingFileHelper();
+            CompletableFuture<HolderLookup.Provider> lookupProvider = event.getLookupProvider();
+
+            generator.addProvider(event.includeServer(), ModLootTableProvider.create(packOutput, lookupProvider));
+
+            generator.addProvider(event.includeClient(), new ModBlockStateProvider(packOutput, existingFileHelper));
+
+            generator.addProvider(event.includeServer(), new ModWorldGenProvider(packOutput, lookupProvider));
+
+        }
+    }
+
+    public static class ModBlockStateProvider extends BlockStateProvider {
+        public ModBlockStateProvider(PackOutput output, ExistingFileHelper exFileHelper) {
+            super(output, BiomeDecorationTest.MODID, exFileHelper);
+        }
+
+        @Override
+        protected void registerStatesAndModels() {
+            blockWithItem(ModBlocks.SILVER_ORE);
+        }
+
+        private void blockWithItem(RegistryObject<Block> blockRegistryObject) {
+            simpleBlockWithItem(blockRegistryObject.get(), cubeAll(blockRegistryObject.get()));
+        }
+    }
+
+    public static class ModBlockLootTables extends BlockLootSubProvider {
+        public ModBlockLootTables() {
+            super(Set.of(), FeatureFlags.REGISTRY.allFlags());
+        }
+
+        @Override
+        protected void generate() {
+            this.dropSelf(ModBlocks.SILVER_ORE.get());
+        }
+
+        @Override
+        protected Iterable<Block> getKnownBlocks() {
+            return ModBlocks.BLOCKS.getEntries().stream().map(RegistryObject::get)::iterator;
+        }
+    }
+
+    public static class ModLootTableProvider {
+        public static LootTableProvider create(PackOutput output, CompletableFuture<HolderLookup.Provider> pRegistries) {
+            return new LootTableProvider(output, Set.of(), List.of(
+                    new LootTableProvider.SubProviderEntry(ModBlockLootTables::new, LootContextParamSets.BLOCK)), pRegistries);
+        }
+    }
+
+    public static class ModWorldGenProvider extends DatapackBuiltinEntriesProvider {
+        public static final RegistrySetBuilder BUILDER = new RegistrySetBuilder()
+                .add(Registries.CONFIGURED_FEATURE, ModConfiguredFeatures::bootstrap)
+                .add(Registries.PLACED_FEATURE, ModPlacedFeatures::bootstrap)
+                .add(ForgeRegistries.Keys.BIOME_MODIFIERS, ModBiomeModifiers::bootstrap);
+
+        public ModWorldGenProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> registries) {
+            super(output, registries, BUILDER, Set.of(BiomeDecorationTest.MODID));
+        }
+    }
+}
